+++ LemonSync/Parser.py
@@ -1,6 +1,35 @@
+# Copyright (c) 2014 LemonStand eCommerce Inc. https://lemonstand.com/
+# All rights reserved.
+#
+# This is free and unencumbered software released into the public domain.
+
+# Anyone is free to copy, modify, publish, use, compile, sell, or
+# distribute this software, either in source code form or as a compiled
+# binary, for any purpose, commercial or non-commercial, and by any
+# means.
+#
+# In jurisdictions that recognize copyright laws, the author or authors
+# of this software dedicate any and all copyright interest in the
+# software to the public domain. We make this dedication for the benefit
+# of the public at large and to the detriment of our heirs and
+# successors. We intend this dedication to be an overt act of
+# relinquishment in perpetuity of all present and future rights to this
+# software under copyright law.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+# IN NO EVENT SHALL THE AUTHORS BE LIABLE FOR ANY CLAIM, DAMAGES OR
+# OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+# ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+# OTHER DEALINGS IN THE SOFTWARE.
+#
+# For more information, please refer to <http://unlicense.org/>
+
 import sys
 import getopt
 from ConfigParser import SafeConfigParser
+from colorama import Fore, Back, Style
 
 sections = 'api', 'dir', 'store'
 
@@ -10,6 +39,6 @@ class Parser (object):
 		parser.optionxform = str
 		found = parser.read(file)
 		if not found:
-			sys.exit('No config file found!')
+			sys.exit(Fore.RED + "No config file found!" + Style.RESET_ALL)
 		for name in sections:
 			self.__dict__.update(parser.items(name))