+++ LemonSync/Utils.py
@@ -1,3 +1,31 @@
+# Copyright (c) 2014 LemonStand eCommerce Inc. https://lemonstand.com/
+# All rights reserved.
+#
+# This is free and unencumbered software released into the public domain.
+
+# Anyone is free to copy, modify, publish, use, compile, sell, or
+# distribute this software, either in source code form or as a compiled
+# binary, for any purpose, commercial or non-commercial, and by any
+# means.
+#
+# In jurisdictions that recognize copyright laws, the author or authors
+# of this software dedicate any and all copyright interest in the
+# software to the public domain. We make this dedication for the benefit
+# of the public at large and to the detriment of our heirs and
+# successors. We intend this dedication to be an overt act of
+# relinquishment in perpetuity of all present and future rights to this
+# software under copyright law.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+# IN NO EVENT SHALL THE AUTHORS BE LIABLE FOR ANY CLAIM, DAMAGES OR
+# OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+# ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+# OTHER DEALINGS IN THE SOFTWARE.
+#
+# For more information, please refer to <http://unlicense.org/>
+
 import sys 
 import time
 import os 
@@ -5,11 +33,15 @@ import boto.s3
 import boto.s3.connection 
 import hashlib
 import shutil
+import requests
+import json
 
 from sys import version_info
 from boto.s3.key import Key
 from watchdog.events import FileSystemEventHandler
 from boto.s3.key import Key
+from colorama import Fore, Back, Style
+from pathtools.patterns import match_path
 
 class Utils ():
 
@@ -25,11 +57,14 @@ class Utils ():
 
 		# Go though each file in s3 and compare it with the local file system
 		for key_val in rs_keys:
-
 			loc_path = key_val.name.replace(path+'/', '')
+			filepath = self.config.watch_dir + loc_path
 
-			# ignore git files
-			if loc_path.startswith(".git"):
+			# Only remove files that are on the watch list
+			if not match_path(filepath,
+				included_patterns=self.config.file_patterns,
+				excluded_patterns=self.config.ignore_patterns,
+				case_sensitive=True):
 				continue
 
 			if not os.path.isfile(self.config.watch_dir + loc_path):
@@ -45,14 +80,17 @@ class Utils ():
 	#  Download changed files from LemonStand and save them locally
 	def sync (self, files):
 		for f in files:
-			# ignore git files
-			if f.startswith(".git"):
-				continue
-				
 			path = os.path.join(self.config.watch_dir, f)
+
+			if not match_path(path,
+				included_patterns=self.config.file_patterns,
+				excluded_patterns=self.config.ignore_patterns,
+				case_sensitive=True):
+				continue
+
 			key = "/".join([self.connection["store"], "themes", self.connection["theme"], f])
 
-			# Split the path into the directoy and the filename
+			# Split the path into the directory and the filename
 			file_dir, file_name = os.path.split(path)
 
 			# Check to see if the file path exists, and create the needed 
@@ -62,20 +100,18 @@ class Utils ():
 
 			if not os.path.isdir(path):
 				# This will create the file if it does not exist
-				f = open(path, 'w+')
+				fi = open(path, 'w+')
 				k = self.connection["bucket"].new_key(key)
-				k.get_contents_to_file(f)
-			# except:
-			# 	sys.exit('Error! Could not create the needed directories.')
+				k.get_contents_to_file(fi)
 
-			print '\033[92m' + '[' + time.strftime("%c") + '] Successfully downloaded ' + key + ''
+			print Fore.GREEN + '[' + time.strftime("%c") + '] Successfully downloaded ' + f + Style.RESET_ALL
 
 		return
 
-	# 
+	# This will overwrite any contents in the remote theme with the contents of the watch_dir
 	def reset_remote (self):
-		print '\033[91m' + 'Are you sure you want to permanlty overwrite your remote theme with the contents of ' + self.config.watch_dir + '?' + '\033[91m'
-		print '\033[91m' + 'Type [Y] to overwrite your remote theme or [q] to quit. Any other key will result in no action being taken.' + '\033[91m'
+		print Back.RED + Fore.WHITE + 'Are you sure you want to permanently overwrite your remote theme with the contents of ' + self.config.watch_dir + ' ?' + Style.RESET_ALL
+		print Fore.RED + 'Type [Y] to overwrite your remote theme or [q] to quit. Any other key will result in no action being taken.' + Style.RESET_ALL
 
 		if version_info[0] > 2:
 			response = input(": ")
@@ -83,12 +119,13 @@ class Utils ():
 			response = raw_input(": ")
 
 		if not response == "Y":
-			sys.exit(1)
+			sys.exit(0)
 
 		self.remove_remote_files()
 
 		# Get the filenames we about to push to s3
 		uploads = []
+		keynames = []
 
 		for (source, dirname, filenames) in os.walk(self.config.watch_dir):
 			for name in filenames:
@@ -99,23 +136,52 @@ class Utils ():
 			keypath = filename.replace(self.config.watch_dir, '')
 			keyname = "/".join([self.connection["store"], "themes", self.connection["theme"], keypath])
 
-			# ignore git files
-			if keypath.startswith(".git"):
+			# Only remove files that are on the watch list
+			if not match_path(filename,
+				included_patterns=self.config.file_patterns,
+				excluded_patterns=self.config.ignore_patterns,
+				case_sensitive=True):
 				continue
+
+			expires = int(time.time())
+			headers = {
+				'Cache-Control': "max-age=" + str(expires) + ", public",
+				'Expires': expires
+			}
 
 			try:
 				k = self.connection["bucket"].new_key(keyname)
-				k.set_contents_from_filename(filename)
-				print '\033[92m' + '[' + time.strftime("%c") + '] Successfully uploaded ' + keyname + ''
+				k.set_contents_from_filename(filename, headers=headers)
+				keynames.append(keypath)
+				print Fore.GREEN + '[' + time.strftime("%c") + '] Successfully uploaded ' + keypath + Style.RESET_ALL
 			except:
-				print '\033[91m' + '[' + time.strftime("%c") + '] Failed to upload ' + keyname + ''
+				print Fore.RED + '[' + time.strftime("%c") + '] Failed to upload ' + keypath + Style.RESET_ALL
 
+		# Notify LS2 that the files have changed
+		try:
+			data = { 'keys': keynames }
+			# Update the resource with LemonStand
+			res = requests.post(
+				self.config.api_host + '/v2/theme/reset', 
+				headers = { 
+					'content-type': 'application/json',
+					'x-store-host': self.config.store_host, 
+					'authorization': self.config.api_access
+				},
+				data=json.dumps(data), 
+				allow_redirects=False
+			)			
+
+			if res.status_code != 200:
+				raise Exception()
+		except:
+			print Fore.RED + '[' + time.strftime("%c") + '] Failed to register local files with LemonStand!' + Style.RESET_ALL
 		return
 
 	def reset_local (self):
 
-		print '\033[91m' + 'Are you sure you want to permanlty overwrite ' + self.config.watch_dir + ' with the remote theme folder?' + '\033[91m'
-		print '\033[91m' + 'Type [Y] to overwrite your local files or [q] to quit. Any other key will result in no action being taken.' + '\033[91m'
+		print Back.RED + Fore.WHITE + 'Are you sure you want to permanently overwrite ' + self.config.watch_dir + ' with the remote theme folder?' + Style.RESET_ALL
+		print Fore.RED + 'Type [Y] to overwrite your local files or [q] to quit. Any other key will result in no action being taken.' + Style.RESET_ALL
 
 		if version_info[0] > 2:
 			response = input(": ")
@@ -123,14 +189,14 @@ class Utils ():
 			response = raw_input(": ")
 
 		if not response == "Y":
-			sys.exit(1)
+			sys.exit(0)
 
 		# Prepare the directory
 		self.remove_local_files(self.config.watch_dir)
 
 		path = os.path.join(self.connection["store"], "themes", self.connection["theme"])
 
-		# retreive the list of keys from the bucket
+		# retrieve the list of keys from the bucket
 		rs_keys = self.connection["bucket"].list(prefix=path)
 
 		new_files = []
@@ -149,7 +215,14 @@ class Utils ():
 		for the_file in os.listdir(directory):
 			file_path = os.path.join(directory, the_file)
 
-			# Do not delete any git files
+			# Only remove files that are on the watch list
+			if not match_path(file_path,
+				included_patterns=self.config.file_patterns,
+				excluded_patterns=self.config.ignore_patterns,
+				case_sensitive=True):
+				continue
+
+			# Just in case the user did not add .git to their list of ignore
 			if the_file.startswith(".git"):
 				continue
 
@@ -159,7 +232,7 @@ class Utils ():
 				elif os.path.isdir(file_path):
 					shutil.rmtree(file_path)
 			except Exception, e:
-				print e
+				print e + Style.RESET_ALL
 		return
 
 	def remove_remote_files (self):
@@ -175,10 +248,10 @@ class Utils ():
 
 	def clean_changes (self, changes):
  		for files in changes:
-			print '--- ' + files
+			print Fore.CYAN + ' --- ' + files + Style.RESET_ALL
 			
-		print '\033[91m' + 'Version mismatch! Do you want to pull in changed files from your remote store?' + '\033[91m'
-		print '\033[91m' + 'Type [Y] to overwrite your local files or [q] to quit. Any other key will result in your local files remaining the same.' + '\033[91m'
+		print Back.RED + Fore.WHITE + 'The above remote files have changed! Do you want to merge these files locally?' + Style.RESET_ALL
+		print Fore.RED + 'Type [Y] to overwrite your local files or [q] to quit. Any other key will result in your local files remaining the same.' + Style.RESET_ALL
 
 		if version_info[0] > 2:
 			response = input(": ")